worked on quizOptions components
